feat: Adapt to bq76920 cell voltage conversion and update submodule
Update src/bq25730_task.rs and src/bq76920_task.rs to handle the converted i32 cell voltage values from the bq76920 submodule.
Adjust temperature printing in src/bq76920_task.rs to show f32 values.
Update bq76920 submodule reference.

+++ src/bq25730_task.rs
@@ -305,7 +305,7 @@ pub async fn bq25730_task(
         }
 
         // Implement battery charge logic based on BQ76920 total battery voltage and final_charge_permission
-        let total_voltage_mv: u16 = bq76920_measurements.core_measurements.cell_voltages.voltages.iter().sum();
+        let total_voltage_mv: i32 = bq76920_measurements.core_measurements.cell_voltages.voltages.iter().sum();
         info!("BQ76920 Total Voltage: {} mV for BQ25730 parameter decision.", total_voltage_mv);
 
         // Define charging parameters
